Konami Cheat Code - Initial
Created Konami inspired code using jQuery where when certain keys are
pressed in a certain order, an alert will pop up.

The keys pressed are added to an array, then checked against an array
storing the correct order the keys should be pressed. If they match, an
alert appears. If not, the second array is emptied and the loop starts
over.